+++ src/Type/src/Exception/TypeNotSupportedException.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace spaceonfire\Type\Exception;
+
+use InvalidArgumentException;
+
+final class TypeNotSupportedException extends InvalidArgumentException
+{
+    /**
+     * TypeNotSupportedException constructor.
+     * @param string $type
+     * @param string|null $typeClass
+     */
+    public function __construct(string $type, ?string $typeClass = null)
+    {
+        parent::__construct(sprintf('Type "%s" is not supported%s', $type, $typeClass ? 'by ' . $typeClass : ''));
+    }
+}