+++ src/Type/src/Factory/ConjunctionTypeFactory.php
@@ -1,0 +1,54 @@
+<?php
+
+declare(strict_types=1);
+
+namespace spaceonfire\Type\Factory;
+
+use spaceonfire\Type\ConjunctionType;
+use spaceonfire\Type\Exception\TypeNotSupportedException;
+use spaceonfire\Type\Type;
+
+final class ConjunctionTypeFactory implements TypeFactoryInterface
+{
+    use TypeFactoryTrait;
+
+    /**
+     * @inheritDoc
+     */
+    public function supports(string $type): bool
+    {
+        if ($this->parent === null) {
+            return false;
+        }
+
+        $parts = explode(ConjunctionType::DELIMITER, $type);
+
+        if (count($parts) < 2) {
+            return false;
+        }
+
+        foreach ($parts as $part) {
+            if (!$this->parent->supports(trim($part))) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function make(string $type): Type
+    {
+        if (!$this->supports($type)) {
+            throw new TypeNotSupportedException($type, ConjunctionType::class);
+        }
+
+        $conjuncts = array_map(function (string $subType): Type {
+            return $this->parent->make(trim($subType));
+        }, explode(ConjunctionType::DELIMITER, $type));
+
+        return new ConjunctionType($conjuncts);
+    }
+}